refactor(bench): extract dashboard helpers to reduce complexity and file size (#247)
